Modify users endpoint
By now, it is considered that users are immediately responsible of every device they have made a proof of. Users endpoint has been modified to adapt to this schema.

First, device addresses for which the user has made a proof are fetched. From that, all proofs of every found device address are returned in a json object.